feat: 정산 상세 조회 [S13P21A310-62]

+++ backend/business/src/main/java/com/udong/backend/dutchpay/dto/CreateDutchpayRequest.java
@@ -19,8 +19,8 @@ public class CreateDutchpayRequest {
     private String note;
 
     @NotNull
-    private Long eventId;
+    private Integer eventId;
 
     @NotEmpty
-    private List<Long> participantUserIds; // users.id 리스트
+    private List<Integer> participantUserIds; // users.id 리스트
 }
